Update ergogen file for v0.2
This is basically rewritten from scratch to work with updated ergogen
syntax.

The most obvious change from v0.1 is the stacked PCB for the central
keys. This should put the central four columns in more comfortable reach
of the index fingers with the hands in a slightly tented posture. It
should also make the home position easier to find and maintain without
homing bumps or ridges.

The front thumb keys remain on the bottom PCB, which should lead to a
natural motion of the thumb and create a better separation between thumb
keys to avoid hitting two at once. It makes the inner-most thumb key
usable.

The bottom PCB still has footprints for all keys and diodes, so it's
still possible to build a flat keyboard with a single PCB.

The pinky keys are rearranged to avoid uncomfortable horizontal
stretching.

The matrix is simplified a bit, and will require changes to the
firmware.

Improved aesthetics with slightly increased space between key blocks and
more consistent spacing and fillets.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>