keycap: Improve choc switch model
Straight and sloping sections in the top and bottom casing. Notches in
the sides of the rim. Larger fillet radii and smoother rendering of the
rounded corners. Thicker walls to align with the LED hole.

Make it a Choc Red switch with a red stem.

Make the cut transparent when rendering the switch+key.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>