keycap: Optimize shape of the outer key slope
Make the outside slope at the rim steeper by inverting the parabola at
the rear of the key cap. This less bulging shape should improve typing
accuracy by minimizing the risk of accidentally hitting adjacent keys
and by making the back and sides of the dish's rim more distinct.

Fixed the calculation of the slope at the front of the key.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>