keycap: Redefine key height parameters
Instead of a key height, specify droop and rise separately. This makes
it explicit, how far the stem is inserted into the key. Only use the
automatic offset calculation to warn if this causes interference between
the key and switch.

In the mantis keyboard model, add parameters for key rise. Make the top
outer pinky keys the same tilt and rise as the home row.

For a proper keywell with 15° tenting, specify tilt2=28, rise2=2.9.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>