keycap: New way to generate constant wall thickness
Minkowski sum creates broken geometry for concave objects, which causes
problems when doing CSG with it. Added an offset parameter to module
fillet_hexagon_cone and tried to generate an object with the walls moved
evenly inwards by that amount. Had to fudge it a bit on the sides.

Used that to create a new offsetkey module.

Also attempted to improve difkey, with questionable success.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>