keycap: v2 keycap STLs
Two variants to make a proper key well with 15° tilt for the home row
and 28° tilt and 3.4mm rise for the outer rows. The upper and lower rows
are rotated 60°, splitting the angle between tenting and the key well:

   tan(tilt) * cos(60) = tan(tent)
   tan(tilt) * sin(60) = tan(well)

Maintaining the tent angle at 15° leads to:

   tan(tilt) = tan(15) / cos(60) <==> tilt = atan(2*tan(15)) = 28.19°

The 28° variant is also raised by 3.4mm. The steeper angle further
raises the center of the dish to about 5mm higher than the 15° variant.
This conforms with the 15° tenting angle for the inner index column.

The excentricity of the 28° key is about 0.9mm more than the 15° key.
Together with the raised and tilted key surface, it should result in
significantly reduced finger movement.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>