Refactoring to rely less on ergogen outlines
Create the case without relying on outlines exported by ergogen. This
provides more flexibility for defining the 3D shape and changing wall
thickness.

Changed side wall thickness and edge fillet to 2mm.

Assume that there will be 1/2 the space around the PCB as there is
around the keycaps. This makes for a tighter fit of the gasket-mounted
PCBs in the case.

Base all resolutions on $fs instead of $fa.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>