Added mounting points and holes
Added mounting points (for threaded inserts) holes for countersunk screws
for mounting the base plate and mezzanine. Also mount the main PCB to the
case near the controller to avoid transmitting unwanted forces to switch
pins when plugging and unplugging the USB.

All the mounting points are accessed from the bottom so that the top of
the case remains perfectly smooth.

TODO: Faciliate movement of the gasket-mounted PCB by adding flex cuts to
the main PCB and splitting the main switch plate into independent halves.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>