Improve and simplify case mounting points
Smoothly integrate main mounting points into the case.

Simplify mounting of the mezzanine: It loosely rests on the rear mounting
points for the base plate and pivots around them during installation. The
front of the trackball holder and case is cut to allow that pivoting
without interference. The two screws holding the mezzanine to the
trackball frame may be optional--it could just be held up by the base
plate.

Added mounting points for the sensor PCB.

Some refactoring and changed case_alpha handling to avoid re-rendering
when alpha is changed.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>