Fix the front of the mezzanine for installation
Cut the entire front of the mezzanine with a rotate_extrude cut, including
the plate itself and the gasket pads. Use the correct outline with the
right offsets and fillets.

Only cut the inside of the case around the trackball holder to make the
trackball holder flush against the case. Outside of that the case it must
still accommodate the PCBs and switch plates.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>