Create saddle-shaped keycap profile
New saddle-shaped keycap profile that acts more like a cylindrical
channel in the home row and more like a tilted spherical dish in the
top and bottom rows. Seems to work well with just one key height and
tilt. I may experiment with a higher version for the inner index keys,
the bottom thumb key and maybe the middle finger column.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>