Update ergogen file for 46-key version
This is a proof of concept of a 46-key Mantis version as proposed by
github user bemyak, based on Mantis v0.3.3. I don't intend to build
this, but maybe it will help bemyak or others develop the idea further.

This updates the ergogen file with 6 additional keys, adds them to the
switch matrix, updates the outlines for PCBs and plates as well as the
positions of mounting holes.

TODO (for whoever picks this up):
- Update controller pin assignment to connect the extra row
- Update header footprints to add an extra row pin
- Update connect_leds.awk to handle the additional LEDs
- Generate PCBs, update footprints, route traces, etc.
- Update QMK firmware to handle the extra keys and LEDs
- Optional: Update the SCAD model

Issue: https://github.com/fxkuehl/mantis/issues/4
Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>