Tweaked keycap profile
Added new type of double saddle using a reverse parabola for positive X.
Similar shape to the cubic double saddle but with the same depth as the
single-saddle and results in slightly lower key cap.

Changed rotation of inner index finger keys to be tilted outwards. This
allows using a tall keycap on the lower-inner position. This was
slightly uncomfortable with spherical dishes due to the sharp edges when
hitting the upper key. But should work better with rounded edges on
saddle-shaped keys.

Made saddle/spherical shape a parameter of the keyboard model.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>