Fixed sensor cut-out
Made the sensor cut-out slightly taller to remove lens interference and
slightly narrower to match the data sheet. It's still shorter than
recommended in the data sheet to allow mor space for traces above the
sensor. This adds a bit of overlap between the sensor package and the
PCB in a place where it should not cause any interference.

Also increased the height of the PCB slightly to mostly compensate for
larger cut-out. This reduces the tolerance at the top and bottom of the
case to about 0.25mm.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>