Optimize connector and and slot positions
Don't use the space under the controller to leave room for a battery and
extra controller pins.

Create some room for cable routing with a larger mezzanine cutout and
larger trackball cutouts in the main PCB and plate that remove the need
for a dedicated cable slot in the middle. Still use a cable slot for the
offset VIK cable. Make a second symmetrical slot in the main plate so
that plate orientation doesn't matter during keyboard assembly.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>