Redo the switch matrix again
Completely redefined the switch matrix from 5x8 to 7x6. This results in
space for two additional keys that can be used for mouse buttons in the
future. Surprisingly this also makes the routing the cleanest of any
Mantis version as the row and column traces are just straight lines on
both PCBs now, with the exception of the bottom keys on the main PCB,
which are appended to the top row.

TODO: Update MCU pin assignments

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>