scad: Add nice!view display
Moved the MCU and rear-edge of the keyboard center up by about 2mm to make
enough room for the display and connector under the top deck. Yay, for
parametric designs. TODO: update the PCBs and plates to match.

Added a 2mm bump on top of the case as a frame around the display. The
display is wedged into the case and sits on top of a lip near the back and
a compliant latch near the front.

The display including frame can be disabled for creating a case for builds
without a display.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>