ergogen: Battery cable management
Cut a groove for the bettery cable into the main PCB foam from the
battery compartment to the cutout under the controller. This should allow
easier battery installation and cable management with minimal
disassembly.

TODO: in KiCad flip the battery connector to face to the back or add an
option to install it eithner way. Add a notch in the PCB edge to allow
the cable to pass through without getting wedged between PCB and case.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>