scad: Add new fillet style for display bump
Renamed "Parabolic" to "Root" and created new "Parabolic" style for the
display bump. "Root" transitions from vertical on the sides to a 30°
angle on the top. "Parabolic" transitions from a 60° angle on the side
to horizontal on the top. This is closer to mimicking the shap of key
caps.

Also played with the bump radii.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>