scad: Make offset_fillet_poly more robust
Detect fillets intersecting with neighbors and truncate them at a
suitable angle. This avoids broken geometry that can also confuse some
version of OpenSCAD.

The strategy depends on whether the direction of adjacent corners is the
same or different. Same direction truncates the fillets at their
intersection point (TODO: support adjacent fillets of different radius).
Different direction truncates the fillets to create a common tangent
(already supports different radii). This does not change the number of
points. Instead the position along the arc is compressed gradually.

Use this to optimize the corner fillet of main key slots to make the
skirts smooth.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>