scad: Separate in/outside xy-fillet of polyhedron
Separate inside and outside fille of polyhedron to get smoother shape on
the bottom of the skirts. Also calculate the corner filler diameter of
the key slots exactly to get the same top diameter on the outside
corners of the key slot fillet as on the inside corners of the top deck.
Preserve the correct fille diameters for the actual key slot cut-out by
separating it into two polyhedra, one for the cut-out and one for
rounding-off the skirt.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>