scad: Optimize skirt thickness and fillets
Let the bottom of the skirt walls be slightly thicker. The raised key
slots only need to cut deep enough to accommodate 3.3mm key travel. Make
Make the offset of the skirt fillet (nearly) 0 so the slop at the bottom
approaches 0. Estimate the z-position of the fillet based on the raised
wall thickness.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>